+++ api/src/middlewares/error.middleware.ts
@@ -2,7 +2,7 @@ import { Context, State, Status } from '@oak/oak';
 import type { Middleware } from '@oak/oak';
 import { APIError, isAPIError } from '../errors/error.ts';
 import { logger } from 'shared/logger.ts';
-import { globalConfig } from 'shared/configManager.ts';
+import { ConfigManager } from 'shared/configManager.ts';
 
 /**
  * Error Handler Middleware function
@@ -18,6 +18,7 @@ export const errorHandler: Middleware = async (
 	try {
 		await next();
 	} catch (err) {
+		const globalConfig = await ConfigManager.globalConfig();
 		if (isAPIError(err)) {
 			const error: APIError = err;
 			const message: string = error.message || 'An error occurred';