+++ bui/src/fresh.gen.ts
@@ -5,8 +5,6 @@
 import * as $_404 from './routes/_404.tsx';
 import * as $_app from './routes/_app.tsx';
 import * as $_middleware from './routes/_middleware.ts';
-import * as $api_joke from './routes/api/joke.ts';
-import * as $greet_name_ from './routes/greet/[name].tsx';
 import * as $index from './routes/index.tsx';
 import * as $Chat from './islands/Chat.tsx';
 import * as $Counter from './islands/Counter.tsx';
@@ -17,8 +15,6 @@ const manifest = {
 		'./routes/_404.tsx': $_404,
 		'./routes/_app.tsx': $_app,
 		'./routes/_middleware.ts': $_middleware,
-		'./routes/api/joke.ts': $api_joke,
-		'./routes/greet/[name].tsx': $greet_name_,
 		'./routes/index.tsx': $index,
 	},
 	islands: {