+++ src/shared/config/configSchema.ts
@@ -1,42 +1,56 @@
 import { simpleMerge } from '@cross/deepmerge';
 
+export interface ApiConfigSchema {
+	anthropicApiKey?: string;
+	openaiApiKey?: string;
+	voyageaiApiKey?: string;
+	environment?: string;
+	apiHostname?: string;
+	apiPort?: number;
+	ignoreLLMRequestCache?: boolean;
+	logFile?: string;
+	logLevel: 'debug' | 'info' | 'warn' | 'error';
+}
+export interface RepoInfoConfigSchema {
+	ctagsAutoGenerate: boolean;
+	ctagsFilePath?: string;
+	tokenLimit?: number;
+}
+export interface ProjectDataConfigSchema {
+	name: string;
+	type: 'local' | 'git';
+	llmGuidelinesFile?: string;
+}
+
 export interface GlobalConfigSchema {
 	myPersonsName?: string;
 	myAssistantsName?: string;
-	api: {
-		anthropicApiKey?: string;
-		openaiApiKey?: string;
-		voyageaiApiKey?: string;
-		environment?: string;
-		apiPort?: number;
-		ignoreLLMRequestCache?: boolean;
-		logFile?: string;
-		logLevel: 'debug' | 'info' | 'warn' | 'error';
-	};
+	noBrowser?: boolean;
+	api: ApiConfigSchema;
 	cli: Record<string, unknown>;
-	repoInfo: {
-		ctagsAutoGenerate: boolean;
-		ctagsFilePath?: string;
-		tokenLimit?: number;
-	};
+	repoInfo: RepoInfoConfigSchema;
 	version: string;
 }
 
 export interface ProjectConfigSchema {
-	project: {
-		name: string;
-		type: 'local' | 'git';
-		llmGuidelinesFile?: string;
-	};
+	myPersonsName?: string;
+	myAssistantsName?: string;
+	noBrowser?: boolean;
+	project: ProjectDataConfigSchema;
+	repoInfo: RepoInfoConfigSchema;
+	api: ApiConfigSchema;
+	cli: Record<string, unknown>;
 }
 
-export interface ConfigSchema extends GlobalConfigSchema, ProjectConfigSchema {}
+export interface FullConfigSchema extends GlobalConfigSchema, ProjectConfigSchema {}
 
 export const defaultGlobalConfig: GlobalConfigSchema = {
 	myPersonsName: Deno.env.get('USER') || 'User',
 	myAssistantsName: 'Claude',
+	noBrowser: false,
 	api: {
 		environment: 'local',
+		apiHostname: 'localhost',
 		apiPort: 3000,
 		ignoreLLMRequestCache: false,
 		logFile: 'api.log',
@@ -51,13 +65,29 @@ export const defaultGlobalConfig: GlobalConfigSchema = {
 };
 
 export const defaultProjectConfig: ProjectConfigSchema = {
+	myPersonsName: Deno.env.get('USER') || 'User',
+	myAssistantsName: 'Claude',
+	noBrowser: false,
+	api: {
+		environment: 'local',
+		apiHostname: 'localhost',
+		apiPort: 3000,
+		ignoreLLMRequestCache: false,
+		logFile: 'api.log',
+		logLevel: 'info',
+	},
+	cli: {},
+	repoInfo: {
+		ctagsAutoGenerate: true,
+		tokenLimit: 1024,
+	},
 	project: {
 		name: Deno.cwd(),
 		type: 'local',
 	},
 };
 
-export function mergeConfigs(...configs: Partial<ConfigSchema>[]): ConfigSchema {
+export function mergeConfigs(...configs: Partial<FullConfigSchema>[]): FullConfigSchema {
 	const mergedConfig = simpleMerge(defaultGlobalConfig as unknown, defaultProjectConfig, ...configs);
-	return mergedConfig as ConfigSchema;
+	return mergedConfig as FullConfigSchema;
 }