+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Refactored Split config handling; global, project, full
+- Added config for apiHostname
+- Changed BUI to also load apiHostname and startDir from URL
+- Fixes for init wizard
 
 
 ## [0.0.12b-beta] - 2024-09-08